Change: Integrated upstream Log/fmt changes

+++ src/GPUSim/MolflowGPUCLI.cpp
@@ -51,6 +51,51 @@
 #include "FlowMPI.h"
 #include "File.h"
 
+class RuntimeStatPrinter {
+    size_t oldHitsNb{0};
+    size_t oldDesNb{0};
+public:
+    RuntimeStatPrinter() = default;
+    RuntimeStatPrinter(size_t n_hits, size_t n_des) {
+        oldHitsNb = n_hits;
+        oldDesNb = n_des;
+    };
+    void PrintHeader() const{
+        // Print Header at the beginning
+        Log::console_msg_master(1, "\n");
+        if(MFMPI::world_size > 1) {
+            Log::console_msg_master(1, "{:<6} ",
+                                    "Node#");
+        }
+        Log::console_msg_master(1, "{:<14} {:<20} {:<20} {:<20} {:<20} {:<20} {:<20}\n",
+                                "Time",
+                                "#Hits (run)", "#Hits (total)","Hit/sec",
+                                "#Des (run)", "#Des (total)","Des/sec");
+        if(MFMPI::world_size > 1) {
+            Log::console_msg_master(1, "{}",std::string(6,'-').c_str());
+        }
+        Log::console_msg_master(1, "{}\n",std::string(14+20+20+20+20+20+20,'-').c_str());
+    }
+    void Print(double elapsedTime, GlobalSimuState& globState, bool printSum=false) const{
+        if(printSum) {
+            Log::console_msg_master(1, "{}\n",std::string(6+14+20+20+20+20+20+20,'=').c_str());
+            Log::console_msg_master(1, "{:<6} ", "x");
+        }
+        else if(MFMPI::world_size > 1) {
+            Log::console_msg(1, "{:<6} ", MFMPI::world_rank);
+        }
+
+        Log::console_msg(1,"{:<14.2f} {:<20} {:<20} {:<20.2f} {:<20} {:<20} {:<20.2f}\n",
+                         elapsedTime,
+                         globState.globalHits.globalHits.nbMCHit - oldHitsNb, globState.globalHits.globalHits.nbMCHit,
+                         (double) (globState.globalHits.globalHits.nbMCHit - oldHitsNb) /
+                         (elapsedTime),
+                         globState.globalHits.globalHits.nbDesorbed - oldDesNb, globState.globalHits.globalHits.nbDesorbed,
+                         (double) (globState.globalHits.globalHits.nbDesorbed - oldDesNb) /
+                         (elapsedTime));
+    }
+};
+
 void printUsageAndExit( const char* argv0 )
 {
     fprintf( stderr, "Usage  : %s [options]\n", argv0 );
@@ -529,7 +574,7 @@ int mainV1(int argc, char **argv) {
                 std::filesystem::remove(fileNameWithZIP);
             }
             catch (const std::exception &e) {
-                Log::console_error("Error compressing to \n%s\nMaybe file is in use:\n%s",fileNameWithZIP.c_str(),e.what());
+                Log::console_error("Error compressing to \n{}\nMaybe file is in use:\n{}",fileNameWithZIP.c_str(),e.what());
             }
         }
         ZipFile::AddFile(fileNameWithZIP, fullOutFile, FileUtils::GetFilename(fullOutFile));
@@ -538,7 +583,7 @@ int mainV1(int argc, char **argv) {
             std::filesystem::remove(fullOutFile);
         }
         catch (const std::exception &e) {
-            Log::console_error("Error removing\n%s\nMaybe file is in use:\n%s",fullOutFile.c_str(),e.what());
+            Log::console_error("Error removing\n{}\nMaybe file is in use:\n{}",fullOutFile.c_str(),e.what());
         }
     }
     /*std::chrono::duration<double,std::milli> elapsed = finish_total - start_total;
@@ -582,7 +627,7 @@ int main(int argc, char** argv) {
     MFMPI::mpi_initialize();
 #endif
 
-    Log::console_msg_master(1, "%s\n", molflowCliLogo);
+    Log::console_msg_master(1, "{}\n", molflowCliLogo);
 
     SimulationManager simManager{MFMPI::world_rank};
     simManager.interactiveMode = true;
@@ -625,14 +670,16 @@ int main(int argc, char** argv) {
     }
     size_t oldHitsNb = globState.globalHits.globalHits.nbMCHit;
     size_t oldDesNb = globState.globalHits.globalHits.nbDesorbed;
-
+    RuntimeStatPrinter printer(oldHitsNb, oldDesNb);
     // Get autosave file name
     std::string autoSave = InitializerGPU::getAutosaveFile();
 
-
     //simManager.ReloadHitBuffer();
     //simManager.IncreasePriority();
-    Log::console_msg_master(1,"[%s] Commencing simulation for %lu seconds from %lu desorptions.\n", Util::getTimepointString().c_str(), Settings::simDuration, globState.globalHits.globalHits.nbDesorbed);
+    if(Settings::simDuration > 0)
+        Log::console_msg_master(1,"[{}] Commencing simulation for {} seconds from {} desorptions.\n", Util::getTimepointString().c_str(), Settings::simDuration, globState.globalHits.globalHits.nbDesorbed);
+    else if(model->otfParams.desorptionLimit > 0)
+        Log::console_msg_master(1,"[{}] Commencing simulation to {} desorptions from {} desorptions.\n", Util::getTimepointString().c_str(), model->otfParams.desorptionLimit, globState.globalHits.globalHits.nbDesorbed);
 
 #if defined(USE_MPI)
     MPI_Barrier(MPI_COMM_WORLD);
@@ -642,8 +689,8 @@ int main(int argc, char** argv) {
         simManager.StartSimulation();
     }
     catch (const std::exception& e) {
-        Log::console_error("[%d] ERROR: Starting simulation: %s\n",MFMPI::world_rank, e.what());
-        Log::console_error("[%d] File folder %s -- %s\n",MFMPI::world_rank,SettingsIO::workPath.c_str(), SettingsIO::workFile.c_str());
+        Log::console_error("[{}] ERROR: Starting simulation: {}\n",MFMPI::world_rank, e.what());
+        Log::console_error("[{}] File folder {} -- {}\n",MFMPI::world_rank,SettingsIO::workPath.c_str(), SettingsIO::workFile.c_str());
 
 #if defined(USE_MPI)
         MPI_Finalize();
@@ -688,7 +735,7 @@ int main(int argc, char** argv) {
                 }
                 writer.SaveSimulationState(outFile, model, globState);
             } catch(std::filesystem::filesystem_error& e) {
-                Log::console_error("Warning: Could not create file: %s\n", e.what());
+                Log::console_error("Warning: Could not create file: {}\n", e.what());
             }
 
             // if there is a next des limit, handle that
@@ -697,40 +744,29 @@ int main(int argc, char** argv) {
                 Settings::desLimit.pop_front();
                 simManager.ForwardOtfParams(&model->otfParams);
                 endCondition = false;
-                Log::console_msg_master(1, " Handling next des limit %z\n", model->otfParams.desorptionLimit);
+                Log::console_msg_master(1, " Handling next des limit {}\n", model->otfParams.desorptionLimit);
 
                 try {
                     ProcessSleep(1000);
                     simManager.StartSimulation();
                 }
                 catch (const std::exception& e) {
-                    Log::console_error("ERROR: Starting simulation: %s\n", e.what());
+                    Log::console_error("ERROR: Starting simulation: {}\n", e.what());
                     endCondition = true;
                 }
             }
         }
         else if(Settings::autoSaveDuration && (uint64_t)(elapsedTime)%Settings::autoSaveDuration==0){ // autosave every x seconds
-            Log::console_msg_master(2,"[%.0lfs] Creating auto save file %s\n", elapsedTime, autoSave.c_str());
+            Log::console_msg_master(2,"[{:.2}s] Creating auto save file {}\n", elapsedTime, autoSave.c_str());
             FlowIO::WriterXML writer;
             writer.SaveSimulationState(autoSave, model, globState);
         }
 
         if(Settings::outputDuration && (uint64_t)(elapsedTime)%Settings::outputDuration==0){ // autosave every x seconds
             if((uint64_t)elapsedTime / Settings::outputDuration <= 1){
-                Log::console_msg_master(1, "\n%-6s %-14s %-20s %-20s %-20s %-20s %-20s %-20s\n",
-                                        "Node#", "Time",
-                                        "#Hits (run)", "#Hits (total)","Hit/sec",
-                                        "#Des (run)", "#Des (total)","Des/sec");
-                Log::console_msg_master(1, "%s\n",std::string(6+14+20+20+20+20+20+20,'-').c_str());
+                printer.PrintHeader();
             }
-            Log::console_msg(1,"%-6d %-14.2lf %-20zu %-20zu %-20.2lf %-20zu %-20zu %-20.2lf\n",
-                             MFMPI::world_rank, elapsedTime,
-                             globState.globalHits.globalHits.nbMCHit - oldHitsNb, globState.globalHits.globalHits.nbMCHit,
-                             (double) (globState.globalHits.globalHits.nbMCHit - oldHitsNb) /
-                             (elapsedTime),
-                             globState.globalHits.globalHits.nbDesorbed - oldDesNb, globState.globalHits.globalHits.nbDesorbed,
-                             (double) (globState.globalHits.globalHits.nbDesorbed - oldDesNb) /
-                             (elapsedTime));
+            printer.Print(elapsedTime, globState);
         }
 
         // Check for potential time end
@@ -745,31 +781,20 @@ int main(int argc, char** argv) {
     simManager.StopSimulation();
     simManager.KillAllSimUnits();
     GatherResults(*model, globState);
-    Log::console_msg(1,"[%d][%s] Simulation finished!\n", MFMPI::world_rank, Util::getTimepointString().c_str());
+    Log::console_msg(1,"[{}][{}] Simulation finished!\n", MFMPI::world_rank, Util::getTimepointString().c_str());
 
 #ifdef USE_MPI
     MPI_Barrier(MPI_COMM_WORLD);
 #endif
-    Log::console_msg_master(1, "\n%-6s %-14s %-20s %-20s %-20s %-20s %-20s %-20s\n",
-                            "Node#", "Time",
-                            "#Hits (run)", "#Hits (total)","Hit/sec",
-                            "#Des (run)", "#Des (total)","Des/sec");
-    Log::console_msg_master(1, "%s\n",std::string(6+14+20+20+20+20+20+20,'-').c_str());
+    printer.PrintHeader();
 #ifdef USE_MPI
     MPI_Barrier(MPI_COMM_WORLD);
 #endif
 
-    //TODO: Send output to master node
+    //TODO: Send output to master node for ordered output
     if(elapsedTime > 1e-4) {
         // Global result print --> TODO: ()
-        Log::console_msg(1,"%-6d %-14.2lf %-20zu %-20zu %-20.2lf %-20zu %-20zu %-20.2lf\n",
-                         MFMPI::world_rank, elapsedTime,
-                         globState.globalHits.globalHits.nbMCHit - oldHitsNb, globState.globalHits.globalHits.nbMCHit,
-                         (double) (globState.globalHits.globalHits.nbMCHit - oldHitsNb) /
-                         (elapsedTime),
-                         globState.globalHits.globalHits.nbDesorbed - oldDesNb, globState.globalHits.globalHits.nbDesorbed,
-                         (double) (globState.globalHits.globalHits.nbDesorbed - oldDesNb) /
-                         (elapsedTime));
+        printer.Print(elapsedTime, globState);
     }
 
 #if defined(USE_MPI)
@@ -789,18 +814,11 @@ int main(int argc, char** argv) {
     if(MFMPI::world_rank != 0){
         return 0;
     }
-    Log::console_msg_master(1, "%s\n",std::string(6+14+20+20+20+20+20+20,'=').c_str());
 #endif //USE_MPI
 
     if(elapsedTime > 1e-4) {
-        Log::console_msg(1,"%-6s %-14.2lf %-20zu %-20zu %-20.2lf %-20zu %-20zu %-20.2lf\n",
-                         "x", elapsedTime,
-                         globState.globalHits.globalHits.nbMCHit - oldHitsNb, globState.globalHits.globalHits.nbMCHit,
-                         (double) (globState.globalHits.globalHits.nbMCHit - oldHitsNb) /
-                         (elapsedTime),
-                         globState.globalHits.globalHits.nbDesorbed - oldDesNb, globState.globalHits.globalHits.nbDesorbed,
-                         (double) (globState.globalHits.globalHits.nbDesorbed - oldDesNb) /
-                         (elapsedTime));
+        if(MFMPI::world_size > 1)
+            printer.Print(elapsedTime, globState, true);
     }
 
     if(MFMPI::world_rank == 0){
@@ -826,9 +844,14 @@ int main(int argc, char** argv) {
         }
         else if(!SettingsIO::overwrite){
             // Copy full file description first, in case outputFile is different
-            if(!SettingsIO::workFile.empty())
-                std::filesystem::copy_file(SettingsIO::workFile, fullOutFile,
-                                           std::filesystem::copy_options::overwrite_existing);
+            if(!SettingsIO::workFile.empty() && std::filesystem::exists(SettingsIO::workFile)){
+                try {
+                    std::filesystem::copy_file(SettingsIO::workFile, fullOutFile,
+                                               std::filesystem::copy_options::overwrite_existing);
+                } catch (std::filesystem::filesystem_error &e) {
+                    Log::console_error("Could not copy file to preserve initial file layout: {}\n", e.what());
+                }
+            }
         }
         FlowIO::WriterXML writer(false, true);
         pugi::xml_document newDoc;
@@ -847,7 +870,7 @@ int main(int argc, char** argv) {
                     std::filesystem::remove(fileNameWithZIP);
                 }
                 catch (std::exception &e) {
-                    Log::console_error("Error compressing to \n%s\nMaybe file is in use:\n%s",fileNameWithZIP.c_str(),e.what());
+                    Log::console_error("Error compressing to \n{}\nMaybe file is in use:\n{}",fileNameWithZIP.c_str(),e.what());
                 }
             }
             ZipFile::AddFile(fileNameWithZIP, fullOutFile, FileUtils::GetFilename(fullOutFile));
@@ -856,7 +879,7 @@ int main(int argc, char** argv) {
                 std::filesystem::remove(fullOutFile);
             }
             catch (std::exception &e) {
-                Log::console_error("Error removing\n%s\nMaybe file is in use:\n%s",fullOutFile.c_str(),e.what());
+                Log::console_error("Error removing\n{}\nMaybe file is in use:\n{}",fullOutFile.c_str(),e.what());
             }
         }
     }
